Delete file if it's there

+++ streaming_generators/scripts/generate_messages.sh
@@ -1,4 +1,4 @@
 #!/usr/bin/env bash
 
-rm messages
-for i in {1..1000}; do echo '{"message_body": "'$(uuidgen)-${i}'"}' >> messages; done
+[ -f messages ] && rm messages
+for i in {1..10000}; do echo '{"message_body": "'$(uuidgen)-${i}'"}' >> messages; done
